Fixed Plant Container Interactions
Issue was that the 'Stretched' mode for the panel and the buttons was somehow getting in the way of the touch event being passed to the button. I reset the position and scale of these on plantcontaineryellow.prefab.
I also realized that in Jon's scene we forgot to add the 'Pointable Canvas Module' to the Event System (now added).